+++ tsetmc/general.py
@@ -175,3 +175,8 @@ class FundType(_StrEnum):
 async def get_funds(type_: FundType | int | str, /) -> _DataFrame:
     j = await _api(f'Fund/GetFunds/{type_}')
     return _DataFrame(j['funds'])
+
+
+async def commodity_funds(*, flow: int | str = '7', top: int | str = '9999'):
+    j = await _api(f'ClosingPrice/GetTradeTop/CommodityFund/{flow}/{top}')
+    return _json_normalize(j['tradeTop'])